requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
pyyaml>=6.0
numpy>=1.24.0
jieba>=0.42.1
